Teensy 3.5 support
Initial commit to allow Teensy 3.5 to work with the
display.  This also adds support for using a Frame buffer
that uses the Pallet code we are experimenting with for
the Teensy 4.0 beta board as well as the Teensy 3.6

This code is starting to work, still needs testing for
SPI1 and SPI2.

Also there is still some issues with the updateDisplayAsync
being offset by I believe a pixel..

This code also includes a hack as the DMA is loosing the CS pin being
held, so the code currently changes the CS pin from using SPI CS code
to use standard digital IO.

ReadRect now look at fifo size.