WIP - T4  Frame buffer can be 16 bits
Right now I have it as a #define in the header file (i.e. you can not change per sketch)....

But T4 now has enough memory possible for full 16 bit frame buffer, so try to build to support it.

Also fixed a few issues, like  some of the write bitmap functions, which specify a specific number of bits,  was not converted to use frame buffer.  Hacked it up, if frame buffer is turned on, call through to the generic function which specifies how many bits per pixel...

Also put code in to allow the tft.begin() to specify which SPI speed to use.