looks for header files using #if __hasinclude(...)
It now looks to see if compiling for T4.1 and the sketch has include extram.h
If so it turns on the using external memory for frame buffer.

Also looks for another header file that if found it inlcudes it as a way for sketches to override the settings, without having to change this library.

Currently the only way I could get this to work is by creating another bogus library where I put the header file such that the Adriuno build process finds it.