Don't clear frame buffer + call frame_complete_callback
With ILI9341_t3n we removed the clearing of the buffer, to allow
people to swap buffers and the like.   Thought it would be good here
as well.

The 	void	setFrameCompleteCB(void (*pcb)(), bool fCallAlsoHalfDone=false);

Was not getting called in our normal case, where the updateScreenAsync code is converting RGB565 into RGB888 as part of the DMA/Interrupt code.